+++ core/dbt/clients/git.py
@@ -6,11 +6,19 @@ from dbt.logger import GLOBAL_LOGGER as logger
 import dbt.exceptions
 
 
-def clone(repo, cwd, dirname=None, remove_git_dir=False, branch=None):
+def _is_commit(revision: str) -> bool:
+    # match SHA-1 git commit
+    return bool(re.match(r"\b[0-9a-f]{40}\b", revision))
+
+
+def clone(repo, cwd, dirname=None, remove_git_dir=False, revision=None):
+    has_revision = revision is not None
+    is_commit = _is_commit(revision or "")
+
     clone_cmd = ['git', 'clone', '--depth', '1']
 
-    if branch is not None:
-        clone_cmd.extend(['--branch', branch])
+    if has_revision and not is_commit:
+        clone_cmd.extend(['--branch', revision])
 
     clone_cmd.append(repo)
 
@@ -31,33 +39,38 @@ def list_tags(cwd):
     return tags
 
 
-def _checkout(cwd, repo, branch):
-    logger.debug('  Checking out branch {}.'.format(branch))
+def _checkout(cwd, repo, revision):
+    logger.debug('  Checking out revision {}.'.format(revision))
 
-    run_cmd(cwd, ['git', 'remote', 'set-branches', 'origin', branch])
-    run_cmd(cwd, ['git', 'fetch', '--tags', '--depth', '1', 'origin', branch])
+    fetch_cmd = ["git", "fetch", "origin", "--depth", "1"]
 
-    tags = list_tags(cwd)
-
-    # Prefer tags to branches if one exists
-    if branch in tags:
-        spec = 'tags/{}'.format(branch)
+    if _is_commit(revision):
+        run_cmd(cwd, fetch_cmd + [revision])
     else:
-        spec = 'origin/{}'.format(branch)
+        run_cmd(cwd, ['git', 'remote', 'set-branches', 'origin', revision])
+        run_cmd(cwd, fetch_cmd + ["--tags", revision])
+
+    if _is_commit(revision):
+        spec = revision
+    # Prefer tags to branches if one exists
+    elif revision in list_tags(cwd):
+        spec = 'tags/{}'.format(revision)
+    else:
+        spec = 'origin/{}'.format(revision)
 
     out, err = run_cmd(cwd, ['git', 'reset', '--hard', spec],
                        env={'LC_ALL': 'C'})
     return out, err
 
 
-def checkout(cwd, repo, branch=None):
-    if branch is None:
-        branch = 'HEAD'
+def checkout(cwd, repo, revision=None):
+    if revision is None:
+        revision = 'HEAD'
     try:
-        return _checkout(cwd, repo, branch)
+        return _checkout(cwd, repo, revision)
     except dbt.exceptions.CommandResultError as exc:
         stderr = exc.stderr.decode('utf-8').strip()
-    dbt.exceptions.bad_package_spec(repo, branch, stderr)
+    dbt.exceptions.bad_package_spec(repo, revision, stderr)
 
 
 def get_current_sha(cwd):
@@ -71,7 +84,7 @@ def remove_remote(cwd):
 
 
 def clone_and_checkout(repo, cwd, dirname=None, remove_git_dir=False,
-                       branch=None):
+                       revision=None):
     exists = None
     try:
         _, err = clone(repo, cwd, dirname=dirname,
@@ -97,7 +110,7 @@ def clone_and_checkout(repo, cwd, dirname=None, remove_git_dir=False,
         logger.debug('Pulling new dependency {}.', directory)
     full_path = os.path.join(cwd, directory)
     start_sha = get_current_sha(full_path)
-    checkout(full_path, repo, branch)
+    checkout(full_path, repo, revision)
     end_sha = get_current_sha(full_path)
     if exists:
         if start_sha == end_sha: