+++ test/integration/039_config_test/test_configs.py
@@ -139,6 +139,7 @@ class TestDisabledConfigs(DBTIntegrationTest):
         return {
             'config-version': 2,
             'data-paths': ['data'],
+            'test-paths': ['tests'],
             'models': {
                 'test': {
                     'enabled': "{{ (target.name == 'default2' | as_bool) }}",
@@ -159,6 +160,12 @@ class TestDisabledConfigs(DBTIntegrationTest):
                     },
                 },
             },
+            'tests': {
+                'test': {
+                    'enabled': "{{ (target.name == 'default2') | as_bool }}",
+                    'severity': 'WARN'
+                },
+            },
         }
 
     @property
@@ -177,12 +184,16 @@ class TestDisabledConfigs(DBTIntegrationTest):
         self.assertEqual(len(results), 0)
         results = self.run_dbt(['run', '--target', 'disabled'], strict=False)
         self.assertEqual(len(results), 0)
+        results = self.run_dbt(['test', '--target', 'disabled'], strict=False)
+        self.assertEqual(len(results), 0)
 
         # has seeds/models - enabled should eval to True because of the target
         results = self.run_dbt(['seed'])
         self.assertEqual(len(results), 1)
         results = self.run_dbt(['run'])
         self.assertEqual(len(results), 2)
+        results = self.run_dbt(['test'], strict=False)
+        self.assertEqual(len(results), 5)
 
 
 class TestUnusedModelConfigs(DBTIntegrationTest):
@@ -195,6 +206,7 @@ class TestUnusedModelConfigs(DBTIntegrationTest):
         return {
             'config-version': 2,
             'data-paths': ['data'],
+            'test-paths': ['does-not-exist'],
             'models': {
                 'test': {
                     'enabled': True,
@@ -207,7 +219,12 @@ class TestUnusedModelConfigs(DBTIntegrationTest):
                 'test': {
                     'enabled': True,
                 }
-            }
+            },
+            'tests': {
+                'test': {
+                    'enabled': True,
+                }
+            },
         }
 
     @property
@@ -222,5 +239,60 @@ class TestUnusedModelConfigs(DBTIntegrationTest):
         self.assertIn('Configuration paths exist', str(exc.exception))
         self.assertIn('- sources.test', str(exc.exception))
         self.assertIn('- models.test', str(exc.exception))
+        self.assertIn('- models.test', str(exc.exception))
 
         self.run_dbt(['seed'], strict=False)
+
+class TestConfigIndivTests(DBTIntegrationTest):
+    @property
+    def schema(self):
+        return "config_039"
+
+    @property
+    def models(self):
+        return "models"
+    
+    @property
+    def project_config(self):
+        return {
+            'config-version': 2,
+            'data-paths': ['data'],
+            'test-paths': ['tests'],
+            'seeds': {
+                'quote_columns': False,
+            },
+            'vars': {
+                'test': {
+                    'seed_name': 'seed',
+                }
+            },
+            'tests': {
+                'test': {
+                    'enabled': True,
+                    'severity': 'WARN'
+                }
+            }
+        }
+
+    @use_profile('postgres')
+    def test_postgres_configuring_individual_tests(self):
+        self.assertEqual(len(self.run_dbt(['seed'])), 1)
+        self.assertEqual(len(self.run_dbt(['run'])), 2)
+        
+        # all tests on (minus sleeper_agent) + WARN
+        self.assertEqual(len(self.run_dbt(['test'], strict=False)), 5)
+        
+        # turn off two of them directly
+        self.assertEqual(len(self.run_dbt(['test', '--vars', '{"enabled_direct": False}'], strict=False)), 3)
+        
+        # turn on sleeper_agent data test directly
+        self.assertEqual(len(self.run_dbt(['test', '--models', 'sleeper_agent', 
+            '--vars', '{"enabled_direct": True}'], strict=False)), 1)
+        
+        # set three to ERROR directly
+        results = self.run_dbt(['test', '--models', 'config.severity:error',
+            '--vars', '{"enabled_direct": True, "severity_direct": "ERROR"}'
+            ], strict=False, expect_pass = False)
+        self.assertEqual(len(results), 2)
+        self.assertEqual(results[0].status, 'fail')
+        self.assertEqual(results[1].status, 'fail')